Refactored out base class for tainted data analyzers
